requirements.txt
numpy
qrcode
Pillow
pyzbar
opencv-python